+++ base/basetypes.h
@@ -14,10 +14,10 @@ typedef int64_t i64;
 
 class SparseInt {
 public:
-    static int buf_size(char byte0);
-    static int val_size(i64 val);
-    static int dump(i32 val, char* buf);
-    static int dump(i64 val, char* buf);
+    static size_t buf_size(char byte0);
+    static size_t val_size(i64 val);
+    static size_t dump(i32 val, char* buf);
+    static size_t dump(i64 val, char* buf);
     static i32 load_i32(const char* buf);
     static i64 load_i64(const char* buf);
 };
@@ -32,7 +32,7 @@ public:
     i32 get() const {
         return val_;
     }
-    int val_size() const {
+    size_t val_size() const {
         return SparseInt::val_size(val_);
     }
 };
@@ -47,7 +47,7 @@ public:
     i64 get() const {
         return val_;
     }
-    int val_size() const {
+    size_t val_size() const {
         return SparseInt::val_size(val_);
     }
 };
